+++ src/test/java/com/nextstep/lotto/stringCalculator/StringCalculatorTest.java
@@ -1,7 +1,6 @@
-package com.nextstep.lotto;
+package com.nextstep.lotto.stringCalculator;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
